First commit for ssh-pageant
Introducing ssh-agent, an ssh authentication agent that just passes the
buck to Pageant (from the PuTTY suite).  This lets you use a single key
store for both PuTTY and Cygwin's ssh commands.

It's fairly crude, but there's enough here to be useful.  Connections
are completely serialized, and it's assumed that the clients are well-
behaved.  99% of the time that's probably fine.