Mirror the exit status of child processes
Most caught signals are basically just treated as exit notifiers,
but when a child exits, copy its exit status so ssh-pageant is more
effective as a command wrapper.

All the exit handling is now refactored a bit, with cleanup_signal as
above, cleanup_warn to print a mandatory message, and cleanup_exit
underlying both to remove the socket file and temporary directory.